Refactor (#10)
* Remove redundant test

* Update field type

* Add iteration method

* Add tests

* Remove py scripts

* Extract 'wrapper' package

* Add methods

* Change return type

* Add test

* Rework

* Improve app docs

+++ outdated/sum_dice_base/measure.go
@@ -4,16 +4,16 @@ import (
 	"github.com/DmitryBogomolov/dicecalc/outdated/dice_roller"
 )
 
-type DistinctRollsCalculator func(dice_roller.DiceRoll) int
+type DistinctRollsCalculator func(dice_roller.DiceRoll) uint64
 
 func MakeDistinctRollsCalculator(roller *dice_roller.DiceRoller) DistinctRollsCalculator {
-	factorials := make([]int, roller.DiceCount())
+	factorials := make([]uint64, roller.DiceCount())
 	factorials[0] = 1
-	for i := 1; i < len((factorials)); i++ {
-		factorials[i] = factorials[(i-1)] * (i + 1)
+	for i := 1; i < len(factorials); i++ {
+		factorials[i] = factorials[(i-1)] * uint64(i+1)
 	}
 
-	return func(roll dice_roller.DiceRoll) int {
+	return func(roll dice_roller.DiceRoll) uint64 {
 		n := len(roll)
 		counts := make(map[int]int)
 		for _, dice := range roll {
@@ -27,8 +27,8 @@ func MakeDistinctRollsCalculator(roller *dice_roller.DiceRoller) DistinctRollsCa
 	}
 }
 
-func CalculateDistinctRolls(rolls []dice_roller.DiceRoll, calculate DistinctRollsCalculator) int {
-	count := 0
+func CalculateDistinctRolls(rolls []dice_roller.DiceRoll, calculate DistinctRollsCalculator) uint64 {
+	count := uint64(0)
 	for _, roll := range rolls {
 		count += calculate(roll)
 	}
