+++ outdated/sum_dice_seq/sum_dice.go
@@ -7,10 +7,10 @@ import (
 )
 
 func CalculateProbabilities(params probabilities.DiceRollParameters) (probabilities.Probabilities, error) {
-	calculateValues := func(k int, roller *dice_roller.DiceRoller) []int {
+	calculateValues := func(k int, roller *dice_roller.DiceRoller) []uint64 {
 		calculate := sum_dice_base.MakeDistinctRollsCalculator(roller)
 		rolls := []dice_roller.DiceRoll{roller.IdxToRoll(0)}
-		result := make([]int, k)
+		result := make([]uint64, k)
 		for i := 0; i < k; i++ {
 			result[i] = sum_dice_base.CalculateDistinctRolls(rolls, calculate)
 			rolls = getNextRolls(rolls, roller)
